refactor mcp actor and utils, use real Actor ID as standby URL
